requirements.txt
pytesseract~=0.3.10
pillow~=9.2.0
pdf2image~=1.16.0
opencv-python~=4.6.0.66